Update 02. expression.js
Write a console-based JavaScript program that calculates and prints the value of the following numerical expression:
(3522 + 52353) * 23 - (2336 * 501 + 23432 - 6743) * 3
Note: you are not allowed to previously calculate the value (for example with Windows Calculator).
Hints and Guidelines
Create a new JavaScript file and name it expression.js. Then we need to write the code which will calculate the numeric expression above and print its value on the console. Pass the expression to the console.log(...) command by writing it within its brackets:
Start the program with [F5] and check whether the result matches: -2275950
Note: Don't forget to enclose your code in the solve() function:
function solve() {
    // your code
}